requirements.txt
PySide6>=6.5.2
psutil>=5.9.5
dnspython>=2.4.2
requests>=2.31.0
scapy>=2.5.0
setproctitle>=1.3.2
APScheduler>=3.10.4
pytz>=2023.3
tzlocal>=5.0.1
bottle>=0.12.25
